+++ packages/material-example/src/app/app.component.ts
@@ -1,0 +1,50 @@
+import { NgIf } from '@angular/common';
+import { Component } from '@angular/core';
+import {
+  HdObscureAddressPipe,
+  HdWalletAdapterDirective,
+} from '@heavy-duty/wallet-adapter-cdk';
+import { HdWalletMultiButtonComponent } from '@heavy-duty/wallet-adapter-material';
+
+@Component({
+  standalone: true,
+  selector: 'hd-root',
+  template: `
+    <main
+      *hdWalletAdapter="
+        let wallet = wallet;
+        let connected = connected;
+        let publicKey = publicKey;
+        let wallets = wallets
+      "
+    >
+      <header>
+        <h1>Wallet Adapter Example (CDK)</h1>
+      </header>
+
+      <section>
+        <div>
+          <p>
+            Wallet:
+            {{ wallet !== null ? wallet.adapter.name : 'None' }}
+          </p>
+
+          <p *ngIf="publicKey">
+            Public Key: {{ publicKey.toBase58() | hdObscureAddress }}
+          </p>
+
+          <p>Status: {{ connected ? 'connected' : 'disconnected' }}</p>
+        </div>
+
+        <hd-wallet-multi-button></hd-wallet-multi-button>
+      </section>
+    </main>
+  `,
+  imports: [
+    NgIf,
+    HdWalletAdapterDirective,
+    HdObscureAddressPipe,
+    HdWalletMultiButtonComponent,
+  ],
+})
+export class AppComponent {}